replace unsigned long long and unsigned long to uint64

+++ arch/riscv/include/put.h
@@ -6,7 +6,7 @@
 void puti(int num);
 int puts(const char *s);
 void print(const char *fmt, ...);
-void putX(unsigned long long num);
+void putX(uint64 num);
 
 #define putchar(c) (*(volatile unsigned char *)0x10000000)=(unsigned char)(c)
