+++ include/put.h
@@ -6,7 +6,7 @@
 void puti(int num);
 int puts(const char *s);
 void print(const char *fmt, ...);
-void putX(unsigned long long num);
+void putX(uint64 num);
 #define UART16550A_DR (volatile unsigned char *)0x10000000
 #define putchar(c) (*(volatile unsigned char *)0x10000000)=(unsigned char)(c)
